requirements.txt
asgiref==3.3.1
backports.zoneinfo==0.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-cors-headers==3.7.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.12.2
drf-writable-nested==0.6.3
gunicorn==20.1.0
idna==2.10
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dotenv==0.17.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
tzdata==2021.5
urllib3==1.26.4
whitenoise==5.2.0
